data-preparation/requirements.txt
boto3==1.34.38
botocore==1.34.38
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
google-api-core==2.16.2
google-api-python-client==2.116.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
oauth2client==4.1.3
pandas==2.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyDrive==1.3.1
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
s3transfer==0.10.0
six==1.16.0
tqdm==4.66.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.0.7
